+++ screens/CampaignScreen/Paragraph.js
